add refactor including sendgrid service
