+++ services/sendgrid/index.js
@@ -1,0 +1,1 @@
+export { sendEmail } from './sendgrid';
